Progress! Most issues from last time have been fixed
The menu links work as expected, the pages load, the queries get the correct data, the json files have been moved to 'src' and proper modification has been made and tested.

So far, so good. Just need to make sure that the ShopItems (complete re-work required) are correctly configured to handle the collections update properly.